♻️ Split test_compass_core_utility.py into per module files
