+++ tests/test_compass_core_util_client.py
@@ -1,0 +1,16 @@
+import pytest
+
+import compass.core as ci
+from compass.core.util import client
+
+
+class TestClient:
+    def test_custom_error(self):
+        # assume underlying client is tested by library -- only test what we add
+        # Given
+        session = client.Client()
+
+        # Then
+        with pytest.raises(ci.CompassNetworkError):
+            # When
+            session.get("https://httpbin.org/delay/5", timeout=0.5)
